new test_solr for esgpull.context.solr.SolrContext
