+++ tests/utils.py
@@ -1,10 +1,14 @@
 from typing import Any, Mapping
 
+import pytest
+
 IPSL_NODE = "esgf-node.ipsl.upmc.fr"
 DRKZ_NODE = "esgf-data.dkrz.de"
 CEDA_NODE = "esgf.ceda.ac.uk"
 ORNL_BRIDGE = "esgf-node.ornl.gov/esgf-1-5-bridge"
 CEDA_STAC = "api.stac.esgf.ceda.ac.uk"
+
+parametrized_index = pytest.mark.parametrize("index", [CEDA_NODE, ORNL_BRIDGE])
 
 
 def dict_equals_ignore(
